+++ beatmap/shared/serializable.ts
@@ -6,6 +6,7 @@ export abstract class Serializable<T extends { [P in keyof T]: T[P] }> implement
       return JSON.stringify(this.toJSON());
    }
    clone<U extends this>(): U {
-      return new (this.constructor as { new (data: T): U })(this.toJSON());
+      // deno-lint-ignore no-explicit-any
+      return new (this.constructor as { new (data: any): U })(this);
    }
 }