+++ save/difficulty.ts
@@ -7,14 +7,30 @@ import { DifficultyDataCheck as V1DifficultyCheck } from '../beatmap/v1/dataChec
 import { DifficultyDataCheck as V2DifficultyCheck } from '../beatmap/v2/dataCheck.ts';
 import { DifficultyDataCheck as V3DifficultyCheck } from '../beatmap/v3/dataCheck.ts';
 import { DifficultyDataCheck as V4DifficultyCheck } from '../beatmap/v4/dataCheck.ts';
+import { Difficulty as V1Difficulty } from '../beatmap/v1/difficulty.ts';
+import { Difficulty as V2Difficulty } from '../beatmap/v2/difficulty.ts';
+import { Difficulty as V3Difficulty } from '../beatmap/v3/difficulty.ts';
+import { Difficulty as V4Difficulty } from '../beatmap/v4/difficulty.ts';
+import { IDifficulty as IV1Difficulty } from '../types/beatmap/v1/difficulty.ts';
+import { IDifficulty as IV2Difficulty } from '../types/beatmap/v2/difficulty.ts';
+import { IDifficulty as IV3Difficulty } from '../types/beatmap/v3/difficulty.ts';
+import { IDifficulty as IV4Difficulty } from '../types/beatmap/v4/difficulty.ts';
 import { IWrapDifficulty } from '../types/beatmap/wrapper/difficulty.ts';
 import { resolve } from '../deps.ts';
 import { writeJSONFile, writeJSONFileSync } from '../utils/_fs.ts';
 import { defaultOptions } from './options.ts';
+import { DataCheck } from '../types/beatmap/shared/dataCheck.ts';
 
 function tag(name: string): string[] {
    return ['save', name];
 }
+
+const dataCheckList: Record<number, Record<string, DataCheck>> = {
+   1: V1DifficultyCheck,
+   2: V2DifficultyCheck,
+   3: V3DifficultyCheck,
+   4: V4DifficultyCheck,
+};
 
 export function _difficulty(
    data: IWrapDifficulty,
@@ -31,6 +47,7 @@ export function _difficulty(
          ...options.dataCheck,
       },
       sort: options.sort ?? defaultOptions.difficulty.sort,
+      write: true,
    };
    if (opt.validate.enabled) {
       logger.tInfo(tag('_difficulty'), 'Validating beatmap');
@@ -63,15 +80,7 @@ export function _difficulty(
 
    if (opt.dataCheck.enabled) {
       logger.tInfo(tag('_difficulty'), 'Checking difficulty data value');
-      const dataCheck = ver === 4
-         ? V4DifficultyCheck
-         : ver === 3
-         ? V3DifficultyCheck
-         : ver === 2
-         ? V2DifficultyCheck
-         : ver === 1
-         ? V1DifficultyCheck
-         : {};
+      const dataCheck = dataCheckList[ver] ?? {};
       deepCheck(
          json,
          dataCheck,
@@ -92,21 +101,46 @@ export function _difficulty(
  */
 export function difficulty(
    data: IWrapDifficulty,
+   options?: ISaveOptionsDifficulty,
+   // deno-lint-ignore no-explicit-any
+): Promise<Record<string, any>>;
+export function difficulty(
+   data: V1Difficulty,
+   options?: ISaveOptionsDifficulty,
+): Promise<IV1Difficulty>;
+export function difficulty(
+   data: V2Difficulty,
+   options?: ISaveOptionsDifficulty,
+): Promise<IV2Difficulty>;
+export function difficulty(
+   data: V3Difficulty,
+   options?: ISaveOptionsDifficulty,
+): Promise<IV3Difficulty>;
+export function difficulty(
+   data: V4Difficulty,
+   options?: ISaveOptionsDifficulty,
+): Promise<IV4Difficulty>;
+export function difficulty(
+   data: IWrapDifficulty,
    options: ISaveOptionsDifficulty = {},
 ) {
    logger.tInfo(tag('difficulty'), 'Async saving difficulty');
-   return writeJSONFile(
-      _difficulty(data, options),
-      resolve(
-         options.directory ??
-            (globals.directory || defaultOptions.difficulty.directory),
-         options.filePath ??
-            (data.filename ||
-               defaultOptions.difficulty.filePath ||
-               'UnnamedDifficulty.dat'),
-      ),
-      options.format,
-   );
+   const json = _difficulty(data, options);
+   if (options.write ?? defaultOptions.difficulty.write) {
+      return writeJSONFile(
+         json,
+         resolve(
+            options.directory ??
+               (globals.directory || defaultOptions.difficulty.directory),
+            options.filePath ??
+               (data.filename ||
+                  defaultOptions.difficulty.filePath ||
+                  'UnnamedDifficulty.dat'),
+         ),
+         options.format,
+      ).then(() => json);
+   }
+   return new Promise(() => json);
 }
 
 /**
@@ -117,19 +151,44 @@ export function difficulty(
  */
 export function difficultySync(
    data: IWrapDifficulty,
+   options?: ISaveOptionsDifficulty,
+   // deno-lint-ignore no-explicit-any
+): Record<string, any>;
+export function difficultySync(
+   data: V1Difficulty,
+   options?: ISaveOptionsDifficulty,
+): IV1Difficulty;
+export function difficultySync(
+   data: V2Difficulty,
+   options?: ISaveOptionsDifficulty,
+): IV2Difficulty;
+export function difficultySync(
+   data: V3Difficulty,
+   options?: ISaveOptionsDifficulty,
+): IV3Difficulty;
+export function difficultySync(
+   data: V4Difficulty,
+   options?: ISaveOptionsDifficulty,
+): IV4Difficulty;
+export function difficultySync(
+   data: IWrapDifficulty,
    options: ISaveOptionsDifficulty = {},
 ) {
    logger.tInfo(tag('difficultySync'), 'Sync saving difficulty');
-   writeJSONFileSync(
-      _difficulty(data, options),
-      resolve(
-         options.directory ??
-            (globals.directory || defaultOptions.difficulty.directory),
-         options.filePath ??
-            (data.filename ||
-               defaultOptions.difficulty.filePath ||
-               'UnnamedDifficulty.dat'),
-      ),
-      options.format,
-   );
+   const json = _difficulty(data, options);
+   if (options.write ?? defaultOptions.difficulty.write) {
+      writeJSONFileSync(
+         json,
+         resolve(
+            options.directory ??
+               (globals.directory || defaultOptions.difficulty.directory),
+            options.filePath ??
+               (data.filename ||
+                  defaultOptions.difficulty.filePath ||
+                  'UnnamedDifficulty.dat'),
+         ),
+         options.format,
+      );
+   }
+   return json;
 }