+++ save/info.ts
@@ -6,14 +6,27 @@ import { deepCheck } from '../beatmap/shared/dataCheck.ts';
 import { InfoDataCheck as V1InfoCheck } from '../beatmap/v1/dataCheck.ts';
 import { InfoDataCheck as V2InfoCheck } from '../beatmap/v2/dataCheck.ts';
 import { InfoDataCheck as V4InfoCheck } from '../beatmap/v4/dataCheck.ts';
+import { Info as V1Info } from '../beatmap/v1/info.ts';
+import { Info as V2Info } from '../beatmap/v2/info.ts';
+import { Info as V4Info } from '../beatmap/v4/info.ts';
+import { IInfo as IV1Info } from '../types/beatmap/v1/info.ts';
+import { IInfo as IV2Info } from '../types/beatmap/v2/info.ts';
+import { IInfo as IV4Info } from '../types/beatmap/v4/info.ts';
 import { IWrapInfo } from '../types/beatmap/wrapper/info.ts';
 import { resolve } from '../deps.ts';
 import { writeJSONFile, writeJSONFileSync } from '../utils/_fs.ts';
 import { defaultOptions } from './options.ts';
+import { DataCheck } from '../types/beatmap/shared/dataCheck.ts';
 
 function tag(name: string): string[] {
    return ['save', name];
 }
+
+const dataCheckList: Record<number, Record<string, DataCheck>> = {
+   1: V1InfoCheck,
+   2: V2InfoCheck,
+   4: V4InfoCheck,
+};
 
 function _info(data: IWrapInfo, options: ISaveOptionsInfo) {
    const opt: Required<ISaveOptionsInfo> = {
@@ -27,6 +40,7 @@ function _info(data: IWrapInfo, options: ISaveOptionsInfo) {
          ...options.dataCheck,
       },
       sort: options.sort ?? defaultOptions.info.sort,
+      write: true,
    };
 
    if (opt.sort) {
@@ -43,13 +57,7 @@ function _info(data: IWrapInfo, options: ISaveOptionsInfo) {
 
    if (opt.dataCheck.enabled) {
       logger.tInfo(tag('_info'), 'Checking info data value');
-      const dataCheck = ver === 4
-         ? V4InfoCheck
-         : ver === 2
-         ? V2InfoCheck
-         : ver === 1
-         ? V1InfoCheck
-         : {};
+      const dataCheck = dataCheckList[ver] ?? {};
       deepCheck(
          json,
          dataCheck,
@@ -68,18 +76,30 @@ function _info(data: IWrapInfo, options: ISaveOptionsInfo) {
  * await save.info(info);
  * ```
  */
+export function info(
+   data: IWrapInfo,
+   options?: ISaveOptionsInfo,
+   // deno-lint-ignore no-explicit-any
+): Promise<Record<string, any>>;
+export function info(data: V1Info, options?: ISaveOptionsInfo): Promise<IV1Info>;
+export function info(data: V2Info, options?: ISaveOptionsInfo): Promise<IV2Info>;
+export function info(data: V4Info, options?: ISaveOptionsInfo): Promise<IV4Info>;
 export function info(data: IWrapInfo, options: ISaveOptionsInfo = {}) {
    logger.tInfo(tag('info'), 'Async saving info');
-   return writeJSONFile(
-      _info(data, options),
-      resolve(
-         options.directory ??
-            (globals.directory || defaultOptions.info.directory),
-         options.filePath ??
-            (data.filename || defaultOptions.info.filePath || 'Info.dat'),
-      ),
-      options.format,
-   );
+   const json = _info(data, options);
+   if (options.write ?? defaultOptions.info.write) {
+      return writeJSONFile(
+         json,
+         resolve(
+            options.directory ??
+               (globals.directory || defaultOptions.info.directory),
+            options.filePath ??
+               (data.filename || defaultOptions.info.filePath || 'Info.dat'),
+         ),
+         options.format,
+      );
+   }
+   return new Promise(() => json);
 }
 
 /**
@@ -88,16 +108,28 @@ export function info(data: IWrapInfo, options: ISaveOptionsInfo = {}) {
  * save.infoSync(info);
  * ```
  */
+export function infoSync(
+   data: IWrapInfo,
+   options?: ISaveOptionsInfo,
+   // deno-lint-ignore no-explicit-any
+): Record<string, any>;
+export function infoSync(data: V1Info, options?: ISaveOptionsInfo): IV1Info;
+export function infoSync(data: V2Info, options?: ISaveOptionsInfo): IV2Info;
+export function infoSync(data: V4Info, options?: ISaveOptionsInfo): IV4Info;
 export function infoSync(data: IWrapInfo, options: ISaveOptionsInfo = {}) {
    logger.tInfo(tag('infoSync'), 'Sync saving info');
-   writeJSONFileSync(
-      _info(data, options),
-      resolve(
-         options.directory ??
-            (globals.directory || defaultOptions.info.directory),
-         options.filePath ??
-            (data.filename || defaultOptions.info.filePath || 'Info.dat'),
-      ),
-      options.format,
-   );
+   const json = _info(data, options);
+   if (options.write ?? defaultOptions.info.write) {
+      writeJSONFileSync(
+         json,
+         resolve(
+            options.directory ??
+               (globals.directory || defaultOptions.info.directory),
+            options.filePath ??
+               (data.filename || defaultOptions.info.filePath || 'Info.dat'),
+         ),
+         options.format,
+      );
+   }
+   return json;
 }