+++ save/lightshow.ts
@@ -5,14 +5,24 @@ import logger from '../logger.ts';
 import { deepCheck } from '../beatmap/shared/dataCheck.ts';
 import { LightshowDataCheck as V3LightshowCheck } from '../beatmap/v3/dataCheck.ts';
 import { LightshowDataCheck as V4LightshowCheck } from '../beatmap/v4/dataCheck.ts';
+import { Lightshow as V3Lightshow } from '../beatmap/v3/lightshow.ts';
+import { Lightshow as V4Lightshow } from '../beatmap/v4/lightshow.ts';
+import { ILightshow as IV3Lightshow } from '../types/beatmap/v3/lightshow.ts';
+import { ILightshow as IV4Lightshow } from '../types/beatmap/v4/lightshow.ts';
 import { IWrapLightshow } from '../types/beatmap/wrapper/lightshow.ts';
 import { resolve } from '../deps.ts';
 import { writeJSONFile, writeJSONFileSync } from '../utils/_fs.ts';
 import { defaultOptions } from './options.ts';
+import { DataCheck } from '../types/beatmap/shared/dataCheck.ts';
 
 function tag(name: string): string[] {
    return ['save', name];
 }
+
+const dataCheckList: Record<number, Record<string, DataCheck>> = {
+   3: V3LightshowCheck,
+   4: V4LightshowCheck,
+};
 
 function _lightshow(data: IWrapLightshow, options: ISaveOptionsLightshow) {
    const opt: Required<ISaveOptionsLightshow> = {
@@ -26,6 +36,7 @@ function _lightshow(data: IWrapLightshow, options: ISaveOptionsLightshow) {
          ...options.dataCheck,
       },
       sort: options.sort ?? defaultOptions.lightshow.sort,
+      write: true,
    };
    if (opt.validate.enabled) {
       logger.tInfo(tag('_lightshow'), 'Validating beatmap');
@@ -58,7 +69,7 @@ function _lightshow(data: IWrapLightshow, options: ISaveOptionsLightshow) {
 
    if (opt.dataCheck.enabled) {
       logger.tInfo(tag('_lightshow'), 'Checking lightshow data value');
-      const dataCheck = ver === 4 ? V4LightshowCheck : ver === 3 ? V3LightshowCheck : {};
+      const dataCheck = dataCheckList[ver] ?? {};
       deepCheck(
          json,
          dataCheck,
@@ -79,21 +90,38 @@ function _lightshow(data: IWrapLightshow, options: ISaveOptionsLightshow) {
  */
 export function lightshow(
    data: IWrapLightshow,
+   options?: ISaveOptionsLightshow,
+   // deno-lint-ignore no-explicit-any
+): Promise<Record<string, any>>;
+export function lightshow(
+   data: V3Lightshow,
+   options?: ISaveOptionsLightshow,
+): Promise<IV3Lightshow>;
+export function lightshow(
+   data: V4Lightshow,
+   options?: ISaveOptionsLightshow,
+): Promise<IV4Lightshow>;
+export function lightshow(
+   data: IWrapLightshow,
    options: ISaveOptionsLightshow = {},
 ) {
    logger.tInfo(tag('lightshow'), 'Async saving lightshow');
-   return writeJSONFile(
-      _lightshow(data, options),
-      resolve(
-         options.directory ??
-            (globals.directory || defaultOptions.lightshow.directory),
-         options.filePath ??
-            (data.filename ||
-               defaultOptions.lightshow.filePath ||
-               'UnnamedLightshow.dat'),
-      ),
-      options.format,
-   );
+   const json = _lightshow(data, options);
+   if (options.write ?? defaultOptions.difficulty.write) {
+      return writeJSONFile(
+         json,
+         resolve(
+            options.directory ??
+               (globals.directory || defaultOptions.lightshow.directory),
+            options.filePath ??
+               (data.filename ||
+                  defaultOptions.lightshow.filePath ||
+                  'UnnamedLightshow.dat'),
+         ),
+         options.format,
+      ).then(() => json);
+   }
+   return new Promise(() => json);
 }
 
 /**
@@ -104,19 +132,36 @@ export function lightshow(
  */
 export function lightshowSync(
    data: IWrapLightshow,
+   options?: ISaveOptionsLightshow,
+   // deno-lint-ignore no-explicit-any
+): Record<string, any>;
+export function lightshowSync(
+   data: V3Lightshow,
+   options?: ISaveOptionsLightshow,
+): IV3Lightshow;
+export function lightshowSync(
+   data: V4Lightshow,
+   options?: ISaveOptionsLightshow,
+): IV4Lightshow;
+export function lightshowSync(
+   data: IWrapLightshow,
    options: ISaveOptionsLightshow = {},
 ) {
    logger.tInfo(tag('lightshowSync'), 'Sync saving lightshow');
-   writeJSONFileSync(
-      _lightshow(data, options),
-      resolve(
-         options.directory ??
-            (globals.directory || defaultOptions.lightshow.directory),
-         options.filePath ??
-            (data.filename ||
-               defaultOptions.lightshow.filePath ||
-               'UnnamedLightshow.dat'),
-      ),
-      options.format,
-   );
+   const json = _lightshow(data, options);
+   if (options.write ?? defaultOptions.difficulty.write) {
+      writeJSONFileSync(
+         json,
+         resolve(
+            options.directory ??
+               (globals.directory || defaultOptions.lightshow.directory),
+            options.filePath ??
+               (data.filename ||
+                  defaultOptions.lightshow.filePath ||
+                  'UnnamedLightshow.dat'),
+         ),
+         options.format,
+      );
+   }
+   return json;
 }